int を unsigned int に変更 #1417
- grab_payload(), grab_payload_limited() の引数
- payloadlen, payload_datalen
- 上記の変更箇所に関連する変数
- SSH1 の圧縮使用中の判定を修正

+++ ttssh2/ttxssh/ssh.h
@@ -396,7 +396,7 @@ typedef struct {
 	   packet data for compression. User data is never streamed through here;
 	   it is compressed directly from the user's buffer. */
 	unsigned char *precompress_outbuf;
-	long precompress_outbuflen;
+	unsigned int precompress_outbuflen;
 	/* this is the length of the packet data, including the type header */
 	long outgoing_packet_len;
 
@@ -404,7 +404,7 @@ typedef struct {
 	   packet data. User data is never streamed through here; it is decompressed
 	   directly to the user's buffer. */
 	unsigned char *postdecompress_inbuf;
-	long postdecompress_inbuflen;
+	unsigned int postdecompress_inbuflen;
 
 	// SSH2 packet format:
 	//   | packet_length(4) | padding_length(1) | payload(n1) | padding(n2) | mac |
@@ -416,10 +416,10 @@ typedef struct {
 	//                                            <----------> pvar->ssh_state.payloadlen
 	//                                                         メッセージタイプを含む、ペイロード全体の長さ
 	unsigned char *payload;
-	long payload_grabbed;
-	long payloadlen;
-	long payload_datastart;
-	long payload_datalen;
+	unsigned int payload_grabbed;
+	unsigned int payloadlen;
+	unsigned int payload_datastart;
+	unsigned int payload_datalen;
 
 	uint32 receiver_sequence_number;
 	uint32 sender_sequence_number;
